+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/ModifyAttributeAction.java
@@ -9,13 +9,17 @@ import com.elmakers.mine.bukkit.api.action.CastContext;
 import com.elmakers.mine.bukkit.api.magic.CasterProperties;
 import com.elmakers.mine.bukkit.api.spell.Spell;
 import com.elmakers.mine.bukkit.api.spell.SpellResult;
+import com.elmakers.mine.bukkit.configuration.SpellParameters;
+import com.elmakers.mine.bukkit.utility.SpellUtils;
 
 public class ModifyAttributeAction extends BaseSpellAction
 {
     private String attribute;
-    private double value;
+    private String originalVariable;
+    private String valueString;
     private boolean bypassUndo;
     private String modifyTarget;
+    private SpellParameters spellParameters;
 
     @Override
     public void prepare(CastContext context, ConfigurationSection parameters)
@@ -24,7 +28,11 @@ public class ModifyAttributeAction extends BaseSpellAction
         modifyTarget = parameters.getString("modify_target", "player");
         bypassUndo = parameters.getBoolean("bypass_undo", false);
         attribute = parameters.getString("attribute");
-        value = parameters.getDouble("value");
+        originalVariable = parameters.getString("original_variable", "x");
+        valueString = parameters.getString("value");
+        if (parameters instanceof SpellParameters) {
+            spellParameters = (SpellParameters)parameters;
+        }
     }
 
     @Override
@@ -37,7 +45,22 @@ public class ModifyAttributeAction extends BaseSpellAction
         if (properties == null) {
             return SpellResult.NO_TARGET;
         }
+        Double value = null;
         Double original = properties.getAttribute(attribute);
+        if (original != null && valueString != null && spellParameters != null) {
+            Double transformedValue = SpellUtils.modifyProperty(original, valueString, originalVariable, spellParameters);
+            if (transformedValue != null) {
+                value = transformedValue;
+            }
+        }
+        if (value == null) {
+            try {
+                value = Double.parseDouble(valueString);
+            } catch (Exception ex) {
+                context.getController().getLogger().warning("Unable to parse value in ModifyAttribute: " + valueString);
+                return SpellResult.FAIL;
+            }
+        }
         if (original != null && original == value) {
             return SpellResult.NO_TARGET;
         }
@@ -55,6 +78,7 @@ public class ModifyAttributeAction extends BaseSpellAction
         parameters.add("attribute");
         parameters.add("value");
         parameters.add("modify_target");
+        parameters.add("original_variable");
     }
 
     @Override