+++ Magic/CHANGELOG.md
@@ -65,6 +65,7 @@
  - The `ModifyBlock` action can now spawn falling blocks of any type, rather than only spawning them from destroyed blocks
  - Add `self_destruct` wand property, for a wand that will be destroyed when held
  - Spell parameters can now be used inside of equations within that spell
+ - The `ModifyAttribute` action now supports an `x` variable holding the current attribute value, just like `ModifyProperties`
 
 ## Changes